Dodanie projection (DTO), poprawka SqlClientRepository, ClientService i ClientControllera

+++ src/main/java/io/github/s19151/repository/ClientRepository.java
@@ -2,5 +2,5 @@ package io.github.s19151.repository;
 
 import io.github.s19151.model.Client;
 
-public interface ClientRepository extends BaseRepository<Client, Integer>{
+public interface ClientRepository extends BaseRepository<Client, Long>{
 }
