+++ src/main/java/io/github/s19151/repository/OrderProductRepository.java
@@ -2,6 +2,6 @@ package io.github.s19151.repository;
 
 import io.github.s19151.model.OrderProduct;
 
-public interface OrderProductRepository extends BaseRepository<OrderProduct, Integer> {
+public interface OrderProductRepository extends BaseRepository<OrderProduct, Long> {
 
 }
