+++ src/main/java/io/github/s19151/repository/OrderRepository.java
@@ -2,6 +2,6 @@ package io.github.s19151.repository;
 
 import io.github.s19151.model.Order;
 
-public interface OrderRepository extends BaseRepository<Order, Integer> {
+public interface OrderRepository extends BaseRepository<Order, Long> {
 
 }
