+++ src/main/java/io/github/s19151/repository/ProductRepository.java
@@ -2,6 +2,6 @@ package io.github.s19151.repository;
 
 import io.github.s19151.model.Product;
 
-public interface ProductRepository extends BaseRepository<Product, Integer> {
+public interface ProductRepository extends BaseRepository<Product, Long> {
 
 }
