+++ src/main/java/io/github/s19151/repository/SqlOrderRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import io.github.s19151.model.Order;
 
 @Repository
-public interface SqlOrderRepository extends OrderRepository, JpaRepository<Order, Integer>{
+public interface SqlOrderRepository extends OrderRepository, JpaRepository<Order, Long>{
 
 }
