+++ src/main/java/io/github/s19151/repository/SqlWorkerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import io.github.s19151.model.Worker;
 
 @Repository
-public interface SqlWorkerRepository extends WorkerRepository, JpaRepository<Worker, Integer>{
+public interface SqlWorkerRepository extends WorkerRepository, JpaRepository<Worker, Long>{
 
 }
