+++ src/main/java/io/github/s19151/repository/WorkerRepository.java
@@ -2,6 +2,6 @@ package io.github.s19151.repository;
 
 import io.github.s19151.model.Worker;
 
-public interface WorkerRepository extends BaseRepository<Worker, Integer>{
+public interface WorkerRepository extends BaseRepository<Worker, Long>{
 
 }
